added cluster stats, removed old files, added new error page, fixed various small bugs, new sidebar navigation, better navigation on most pages

+++ neocortex/cache_vmware.py
@@ -25,6 +25,10 @@ def run(helper, options):
 	dcs = {}
 	clusters = {}
 
+	skip_vms = False
+	if options is not None and 'skip_vms' in options and options['skip_vms'] == True:
+		skip_vms = True
+
 	## For each vCenter that appears in the configuration
 	for key in helper.config['VMWARE'].keys():
 		# Get the hostname
@@ -35,65 +39,67 @@ def run(helper, options):
 		content = si.RetrieveContent()
 		helper.end_event(description="Connected to instance " + instance['hostname'])
 
-		vm_properties = ["name", "config.uuid", "config.hardware.numCPU",
-		                 "config.hardware.memoryMB", "runtime.powerState",
-		                 "config.guestFullName", "config.guestId",
-		                 "config.version", "guest.hostName", "guest.ipAddress", 
-		                 "config.annotation", "resourcePool", "guest.toolsRunningStatus", 
-				 "guest.toolsVersionStatus2", "config.template"]
+		# If the task options don't say to skip VM information
+		if not skip_vms:
+			vm_properties = ["name", "config.uuid", "config.hardware.numCPU",
+					 "config.hardware.memoryMB", "runtime.powerState",
+					 "config.guestFullName", "config.guestId",
+					 "config.version", "guest.hostName", "guest.ipAddress", 
+					 "config.annotation", "resourcePool", "guest.toolsRunningStatus", 
+					 "guest.toolsVersionStatus2", "config.template"]
 
-		## List VMs ##########
-		helper.event("vmware_cache_vm", "Downloading virtual machine information from " + instance['hostname'])
+			## List VMs ##########
+			helper.event("vmware_cache_vm", "Downloading virtual machine information from " + instance['hostname'])
 
-		# Get the root of the VMware containers, and search for virtual machines
-		root_folder = si.content.rootFolder
-		view = helper.lib.vmware_get_container_view(si, obj_type=[vim.VirtualMachine])
+			# Get the root of the VMware containers, and search for virtual machines
+			root_folder = si.content.rootFolder
+			view = helper.lib.vmware_get_container_view(si, obj_type=[vim.VirtualMachine])
 
-		# Collect a subset of data on all VMs
-		vm_data_proxy = helper.lib.vmware_collect_properties(si, view_ref=view,
-		                                  obj_type=vim.VirtualMachine,
-		                                  path_set=vm_properties,
-		                                  include_mors=True)
+			# Collect a subset of data on all VMs
+			vm_data_proxy = helper.lib.vmware_collect_properties(si, view_ref=view,
+							  obj_type=vim.VirtualMachine,
+							  path_set=vm_properties,
+							  include_mors=True)
 
-		# Start a vm_data section for this vCenter
-		vm_data[key] = {}
+			# Start a vm_data section for this vCenter
+			vm_data[key] = {}
 
-		# Copy data from the proxy that pyVmomi returns to a non-proxied dictionary
-		for vm in vm_data_proxy:
-			# Get the managed object reference
-			moId = vm['obj']._moId
+			# Copy data from the proxy that pyVmomi returns to a non-proxied dictionary
+			for vm in vm_data_proxy:
+				# Get the managed object reference
+				moId = vm['obj']._moId
 
-			# Start a vm_data section for this vm
-			vm_data[key][moId] = {}
+				# Start a vm_data section for this vm
+				vm_data[key][moId] = {}
 
-			# Put in blank strings for data we don't have
-			for attr in ['config.annotation', 'guest.hostName', 'guest.ipAddress']:
-				if attr not in vm:
-					vm[attr] = ''
+				# Put in blank strings for data we don't have
+				for attr in ['config.annotation', 'guest.hostName', 'guest.ipAddress']:
+					if attr not in vm:
+						vm[attr] = ''
 
-			# Put in the resource pool name rather than a Managed Object
-			if 'resourcePool' in vm:
-				vm_data[key][moId]['cluster'] = vm['resourcePool'].owner.name
-			else:
-				vm_data[key][moId]['cluster'] = "None"
+				# Put in the resource pool name rather than a Managed Object
+				if 'resourcePool' in vm:
+					vm_data[key][moId]['cluster'] = vm['resourcePool'].owner.name
+				else:
+					vm_data[key][moId]['cluster'] = "None"
 
-			# Store moId
-			vm_data[key][moId]['_moId'] = moId
+				# Store moId
+				vm_data[key][moId]['_moId'] = moId
 
-			# Store attributes for the VM
-			for attr in ['name', 'config.uuid', 'config.hardware.numCPU', 'config.hardware.memoryMB', 'runtime.powerState', 'config.guestFullName', 'config.guestId', 'config.version', 'guest.hostName', 'guest.ipAddress', 'config.annotation', 'guest.toolsRunningStatus', 'guest.toolsVersionStatus2', 'config.template']:
-				vm_data[key][moId][attr] = vm[attr]
+				# Store attributes for the VM
+				for attr in ['name', 'config.uuid', 'config.hardware.numCPU', 'config.hardware.memoryMB', 'runtime.powerState', 'config.guestFullName', 'config.guestId', 'config.version', 'guest.hostName', 'guest.ipAddress', 'config.annotation', 'guest.toolsRunningStatus', 'guest.toolsVersionStatus2', 'config.template']:
+					vm_data[key][moId][attr] = vm[attr]
 
-		# Finish the event
-		helper.end_event(description="Downloaded virtual machine information for " + instance['hostname'])
-	
+			# Finish the event
+			helper.end_event(description="Downloaded virtual machine information for " + instance['hostname'])
+		
 		## List DCs ##########
 		helper.event("vmware_cache_dc", "Downloading datacenter information from " + instance['hostname'])
 		dcs[key] = helper.lib.vmware_get_objects(content, [vim.Datacenter])
 		helper.end_event(description="Downloaded datacenter information for " + instance['hostname'])
 
 		## List clusters ##########
-		helper.event("vmware_cache_dc", "Downloading cluster information from " + instance['hostname'])
+		helper.event("vmware_cache_cluster", "Downloading cluster information from " + instance['hostname'])
 		clusters[key] = helper.lib.vmware_get_objects(content, [vim.ClusterComputeResource])
 		helper.end_event(description="Downloaded cluster information from " + instance['hostname'])
 
@@ -105,7 +111,8 @@ def run(helper, options):
 	curd.execute("START TRANSACTION")
 	curd.execute("DELETE FROM `vmware_cache_clusters`")
 	curd.execute("DELETE FROM `vmware_cache_datacenters`")
-	curd.execute("DELETE FROM `vmware_cache_vm`")
+	if not skip_vms:
+		curd.execute("DELETE FROM `vmware_cache_vm`")
 	helper.end_event(description="Deleted existing cache")
 
 	## For each vCenter that appears in the configuration
@@ -116,13 +123,15 @@ def run(helper, options):
 		# Start the event
 		helper.event("vmware_cache_data", "Caching downloaded data from " + instance['hostname'])
 
-		# For each VM
-		for moId in vm_data[key]:
-			# Get the VM
-			vm = vm_data[key][moId]
+		# If the task options doesn't say to skip VM information
+		if not skip_vms:
+			# For each VM
+			for moId in vm_data[key]:
+				# Get the VM
+				vm = vm_data[key][moId]
 
-			# Put the VM in the database
-			curd.execute("INSERT INTO `vmware_cache_vm` (`id`, `vcenter`, `name`, `uuid`, `numCPU`, `memoryMB`, `powerState`, `guestFullName`, `guestId`, `hwVersion`, `hostname`, `ipaddr`, `annotation`, `cluster`, `toolsRunningStatus`, `toolsVersionStatus`, `template`) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)", (vm['_moId'], instance['hostname'], vm['name'], vm['config.uuid'], vm['config.hardware.numCPU'], vm['config.hardware.memoryMB'], vm['runtime.powerState'], vm['config.guestFullName'], vm['config.guestId'], vm['config.version'], vm['guest.hostName'], vm['guest.ipAddress'], vm['config.annotation'], vm['cluster'], vm['guest.toolsRunningStatus'], vm['guest.toolsVersionStatus2'], vm['config.template']))
+				# Put the VM in the database
+				curd.execute("INSERT INTO `vmware_cache_vm` (`id`, `vcenter`, `name`, `uuid`, `numCPU`, `memoryMB`, `powerState`, `guestFullName`, `guestId`, `hwVersion`, `hostname`, `ipaddr`, `annotation`, `cluster`, `toolsRunningStatus`, `toolsVersionStatus`, `template`) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)", (vm['_moId'], instance['hostname'], vm['name'], vm['config.uuid'], vm['config.hardware.numCPU'], vm['config.hardware.memoryMB'], vm['runtime.powerState'], vm['config.guestFullName'], vm['config.guestId'], vm['config.version'], vm['guest.hostName'], vm['guest.ipAddress'], vm['config.annotation'], vm['cluster'], vm['guest.toolsRunningStatus'], vm['guest.toolsVersionStatus2'], vm['config.template']))
 
 		# For each Data Center, put it in the database
 		for dc in dcs[key]:
@@ -134,7 +143,7 @@ def run(helper, options):
 			parent = cluster.parent
 			found = False
 
-			# Loop until we find something
+			# Loop until we find the data center
 			while not found:
 				try:
 					# If we've got a Datacenter object as the parent
@@ -149,8 +158,22 @@ def run(helper, options):
 					clusterdc = "Unknown"
 					break
 
+			total_ram   = 0
+			total_cores = 0
+			total_hz    = 0
+			total_cpu_usage = 0
+			total_ram_usage = 0
+			host_count = len(cluster.host)
+
+			for host in cluster.host:
+				total_ram = total_ram + host.hardware.memorySize
+				total_cores = total_cores + host.hardware.cpuInfo.numCpuCores
+				total_hz = total_hz + host.hardware.cpuInfo.hz
+				total_cpu_usage = total_cpu_usage + host.summary.quickStats.overallCpuUsage
+				total_ram_usage = total_ram_usage + host.summary.quickStats.overallMemoryUsage
+
 			# Put the cluster in the database
-			curd.execute("INSERT INTO `vmware_cache_clusters` (`id`, `name`, `vcenter`, `did`) VALUES (%s, %s, %s, %s)", (cluster._moId, cluster.name, instance['hostname'], clusterdc))
+			curd.execute("INSERT INTO `vmware_cache_clusters` (`id`, `name`, `vcenter`, `did`, `ram`, `cores`, `cpuhz`, `cpu_usage`, `ram_usage`, `hosts`) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s)", (cluster._moId, cluster.name, instance['hostname'], clusterdc, total_ram, total_cores, total_hz, total_cpu_usage, total_ram_usage, host_count))
 
 		helper.end_event(description="Cached information from " + instance['hostname'])
